+++ pmocr.sh
@@ -4,7 +4,7 @@ PROGRAM="pmocr" # Automatic OCR service that monitors a directory and launches a
 AUTHOR="(C) 2015-2016 by Orsiris de Jong"
 CONTACT="http://www.netpower.fr - ozy@netpower.fr"
 PROGRAM_VERSION=1.4.2
-PROGRAM_BUILD=2016081503
+PROGRAM_BUILD=2016081504
 
 ## Debug parameter for service
 _DEBUG=no
@@ -96,7 +96,7 @@ _LOGGER_PREFIX="date"
 
 #### MINIMAL-FUNCTION-SET BEGIN ####
 
-## FUNC_BUILD=2016081501
+## FUNC_BUILD=2016081502
 ## BEGIN Generic functions for osync & obackup written in 2013-2016 by Orsiris de Jong - http://www.netpower.fr - ozy@netpower.fr
 
 ## type -p does not work on platforms other than linux (bash). If if does not work, always assume output is not a zero exitcode
@@ -655,6 +655,7 @@ function WaitForTaskCompletion {
 	local caller_name="${4}" # Who called this function
 	local exit_on_error="${5:-false}" # Should the function exit on subprocess errors
 	local counting="${6:-true}" # Count time since function launch if true, script launch if false
+	local keep_logging="${7}" # Log a standby message every X seconds. Set to zero to disable logging
 
 
 	local soft_alert=0 # Does a soft alert need to be triggered, if yes, send an alert once
@@ -673,8 +674,6 @@ function WaitForTaskCompletion {
 	pidCount=${#pidsArray[@]}
 
 	WAIT_FOR_TASK_COMPLETION=""
-
-	#TODO: need to find a way to properly handle processes in unterruptible sleep state
 
 	while [ ${#pidsArray[@]} -gt 0 ]; do
 		newPidsArray=()
@@ -708,10 +707,12 @@ function WaitForTaskCompletion {
 			exec_time=$SECONDS
 		fi
 
-		if [ $((($exec_time + 1) % $KEEP_LOGGING)) -eq 0 ]; then
-			if [ $log_ttime -ne $exec_time ]; then
-				log_ttime=$exec_time
-				Logger "Current tasks still running with pids [$(joinString , ${pidsArray[@]})]." "NOTICE"
+		if [ $keep_logging -ne 0 ]; then
+			if [ $((($exec_time + 1) % $keep_logging)) -eq 0 ]; then
+				if [ $log_ttime -ne $exec_time ]; then # Fix when sleep time lower than 1s
+					log_ttime=$exec_time
+					Logger "Current tasks still running with pids [$(joinString , ${pidsArray[@]})]." "NOTICE"
+				fi
 			fi
 		fi
 
@@ -938,7 +939,7 @@ function OCR_service {
 	while true
 	do
 		inotifywait --exclude "(.*)$FILENAME_SUFFIX$fileExtension" -qq -r -e create "$directoryToProcess" &
-		WaitForTaskCompletion $! 0 0 ${FUNCNAME[0]} false true
+		WaitForTaskCompletion $! 0 0 ${FUNCNAME[0]} false true 0
 		sleep $WAIT_TIME
 		OCR "$directoryToProcess" "$fileExtension" "$ocrEngineArgs" "$csvHack"
 	done